fix: place Prometheus types in foundation, keep resolution const in ui-core
Addresses review feedback on #9046: Prometheus query-range types are
cross-feature wire shapes (also used by gpuaas), so they belong in
foundation per package-topology. Keep PROMETHEUS_REQUEST_RESOLUTION in
ui-core (no duplicate). TrustyAI/bias types stay in model-serving/shared.
