+++ frontend/src/__mocks__/mockPrometheusBias.ts
@@ -1,6 +1,8 @@
 /* eslint-disable camelcase */
-import type { PrometheusQueryRangeResponseDataResult } from '@odh-dashboard/model-serving/shared/types';
-import { PrometheusQueryRangeResponse } from '#~/types';
+import type {
+  PrometheusQueryRangeResponse,
+  PrometheusQueryRangeResponseDataResult,
+} from '@odh-dashboard/foundation';
 
 type MockPrometheusServingType = {
   result?: PrometheusQueryRangeResponseDataResult[];
